[ADD]Calculator - InquiryService, CalculateService 추출
