+++ src/main/java/com/bona/javacalculator/service/MakePostfixService.java
@@ -1,41 +1,8 @@
 package com.bona.javacalculator.service;
 
-import org.springframework.stereotype.Service;
 import java.util.Stack;
 
-
-public class CalService {
-
-
-//    private static final char WHITE_SPACE = ' ';
-
-    //숫자,연산자 구분
-    public Double calculateStr(String input) {
-        //숫자만 저장
-        Stack<Object> numbers = new Stack<>();
-        char c = ' ';
-        double num1 = 0;
-        double num2 = 0;
-        for (int i = 0; i < input.length(); i++) {
-            c = input.charAt(i);
-            if (Character.isDigit(c)) {
-                numbers.push(c);
-            } else {
-                num2 = Double.parseDouble(numbers.pop().toString());
-                num1 = Double.parseDouble(numbers.pop().toString());
-                //연산자면 계산 후 스택에 저장
-//                calculateReal(num1, num2, c,numbers);
-                switch (c) {
-                    case '+' -> numbers.push(num1 + num2);
-                    case '-' -> numbers.push(num1 - num2);
-                    case '*' -> numbers.push(num1 * num2);
-                    case '/' -> numbers.push(num1 / num2);
-                }
-            }
-        }
-        return Double.valueOf(numbers.pop().toString());
-    }
-
+public class MakePostfixService {
     //후위 표기로 변환
     public String convPostfix(String input) {
         char c = ' ';
